tracecontext/orchestrator/requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain-anthropic
psycopg2-binary
redis
pgvector
python-dotenv
